HeroBonus: use 64 bit for tree counter (to be more robust)

+++ lib/HeroBonus.h
@@ -781,7 +781,7 @@ private:
 	static const bool cachingEnabled;
 	mutable BonusList cachedBonuses;
 	mutable int64_t cachedLast;
-	static std::atomic<int32_t> treeChanged;
+	static std::atomic<int64_t> treeChanged;
 
 	// Setting a value to cachingStr before getting any bonuses caches the result for later requests.
 	// This string needs to be unique, that's why it has to be setted in the following manner:
